created spanner/options.h
