+++ google/cloud/spanner/CMakeLists.txt
@@ -112,6 +112,7 @@ add_library(
     mutations.h
     numeric.cc
     numeric.h
+    options.h
     partition_options.cc
     partition_options.h
     partitioned_dml_result.h
